third_party/WebKit/LayoutTests/fast/dom/HTMLDialogElement/form-method-dialog-expected.txt

Tests form submission with method=dialog

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


clicking outer-submit-empty-string
PASS outer; dialog.open is false
PASS outer; dialog.returnValue is expectedValue
PASS inner; dialog.open is true
PASS inner; dialog.returnValue is "init"
PASS host; dialog.open is true
PASS host; dialog.returnValue is "init"

clicking outer-submit-no-value
PASS outer; dialog.open is false
PASS outer; dialog.returnValue is "init"
PASS inner; dialog.open is true
PASS inner; dialog.returnValue is "init"
PASS host; dialog.open is true
PASS host; dialog.returnValue is "init"

clicking inner-submit-yes
PASS outer; dialog.open is true
PASS outer; dialog.returnValue is "init"
PASS inner; dialog.open is false
PASS inner; dialog.returnValue is expectedValue
PASS host; dialog.open is true
PASS host; dialog.returnValue is "init"

clicking inner-submit-no
PASS outer; dialog.open is true
PASS outer; dialog.returnValue is "init"
PASS inner; dialog.open is false
PASS inner; dialog.returnValue is expectedValue
PASS host; dialog.open is true
PASS host; dialog.returnValue is "init"

clicking no-dialog-ancestor-1
PASS outer; dialog.open is true
PASS outer; dialog.returnValue is "init"
PASS inner; dialog.open is true
PASS inner; dialog.returnValue is "init"
PASS host; dialog.open is true
PASS host; dialog.returnValue is "init"

clicking no-dialog-ancestor-2
PASS outer; dialog.open is true
PASS outer; dialog.returnValue is "init"
PASS inner; dialog.open is true
PASS inner; dialog.returnValue is "init"
PASS host; dialog.open is true
PASS host; dialog.returnValue is "init"

clicking host-submit-yes
PASS outer; dialog.open is true
PASS outer; dialog.returnValue is "init"
PASS inner; dialog.open is true
PASS inner; dialog.returnValue is "init"
PASS host; dialog.open is false
PASS host; dialog.returnValue is expectedValue

clicking a button in a closed dialog
PASS dialog.open is false
PASS dialog.returnValue is "init"
PASS successfullyParsed is true

TEST COMPLETE
  
